New commits to the book and author service.

+++ src/main/java/za/ac/cput/service/IBookService.java
@@ -4,9 +4,9 @@ import za.ac.cput.domain.Book;
 
 import java.util.List;
 
-public interface IBookService extends IService<Book, Long>{
+public interface IBookService extends IService<Book, Integer>{
 
-    void delete(Long bookID);
+    void delete(Integer bookID);
     List<Book> getall();
 
 }
